Add a docker_data dir and improve caching
Rather than relying on things outside the project dir, have a
data dir with all the caches, model files etc, and exclude them from
git and the docker context with the .ignore files.

Also move the install step to last and put it in its own step, so it
doesn't have to rebuild all the slow conda stuff every time there's
a change to a file in the project.